Rename PcapNGFileReader to PcapNGPacketReader
