+++ 2-test/networking.files.tests/PacketReaderFactoryTests/PacketReaderFactory_Create_FilePath_Tests.cs
@@ -12,7 +12,7 @@ namespace Networking.Files.Tests.PacketReaderFactoryTests
         public static List<Object[]> Data => new List<Object[]>
         {
             new Object[] { "pcap.pcap", typeof(PcapPacketReader)},
-            new Object[] { "pcapng.pcapng", typeof(PcapNGFileReader)},
+            new Object[] { "pcapng.pcapng", typeof(PcapNGPacketReader)},
         };
 
         [Fact]
