+++ 2-test/networking.files.tests/PacketReaderFactoryTests/PacketReaderFactory_Create_Stream_Tests.cs
@@ -16,7 +16,7 @@ namespace Networking.Files.Tests.PacketReaderFactoryTests
             new Object[] { new Byte[] { 0xD4, 0xC3, 0xB2, 0xA1 }, typeof(PcapPacketReader)},
             new Object[] { new Byte[] { 0xA1, 0xB2, 0x3C, 0x4D }, typeof(PcapPacketReader)},
             new Object[] { new Byte[] { 0x4D, 0x3C, 0xB2, 0xA1 }, typeof(PcapPacketReader)},
-            new Object[] { new Byte[] { 0x0A, 0x0D, 0x0D, 0x0A }, typeof(PcapNGFileReader)},
+            new Object[] { new Byte[] { 0x0A, 0x0D, 0x0D, 0x0A }, typeof(PcapNGPacketReader)},
         };
 
         [Fact]
