+++ 2-test/networking.files.tests/networking.files.tests.csproj
@@ -21,7 +21,7 @@
 
   <ItemGroup>
     <EmbeddedResource Include="PcapTests\PcapPacketReaderTests\pcap.pcap" />
-    <EmbeddedResource Include="PcapNGTests\PcapNGFileReaderTests\1-shb.11-idb.64-epb.1-nrb.11-isb.pcapng" />
+    <EmbeddedResource Include="PcapNGTests\PcapNGPacketReaderTests\1-shb.11-idb.64-epb.1-nrb.11-isb.pcapng" />
   </ItemGroup>
 
   <ItemGroup>
